+++ src/main/java/codesquad/fineants/domain/kis/service/KisService.java
@@ -39,9 +39,9 @@ import codesquad.fineants.domain.stock_target_price.domain.entity.StockTargetPri
 import codesquad.fineants.domain.stock_target_price.event.publisher.StockTargetPricePublisher;
 import codesquad.fineants.domain.stock_target_price.repository.StockTargetPriceRepository;
 import codesquad.fineants.global.common.delay.DelayManager;
-import codesquad.fineants.global.errors.exception.ExpiredKisAccessTokenException;
-import codesquad.fineants.global.errors.exception.KisException;
-import codesquad.fineants.global.errors.exception.RequestLimitExceededException;
+import codesquad.fineants.global.errors.exception.kis.ExpiredAccessTokenKisException;
+import codesquad.fineants.global.errors.exception.kis.KisException;
+import codesquad.fineants.global.errors.exception.kis.RequestLimitExceededKisException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
@@ -93,14 +93,14 @@ public class KisService {
 				.doOnSuccess(kisCurrentPrice -> log.debug("reload stock current price {}", kisCurrentPrice))
 				.onErrorResume(throwable -> {
 					log.error(throwable.getMessage());
-					if (throwable instanceof ExpiredKisAccessTokenException) {
+					if (throwable instanceof ExpiredAccessTokenKisException) {
 						return Mono.empty();
 					} else {
 						return Mono.error(throwable);
 					}
 				})
 				.retryWhen(Retry.fixedDelay(3, delayManager.fixedDelay())
-					.filter(RequestLimitExceededException.class::isInstance))
+					.filter(RequestLimitExceededKisException.class::isInstance))
 				.blockOptional(delayManager.timeout())
 				.orElse(null);
 			if (price != null) {
