+++ src/main/java/codesquad/fineants/global/errors/exception/kis/ExpiredAccessTokenKisException.java
@@ -1,0 +1,8 @@
+package codesquad.fineants.global.errors.exception.kis;
+
+public class ExpiredAccessTokenKisException extends KisException {
+
+	public ExpiredAccessTokenKisException(String returnCode, String messageCode, String message) {
+		super(returnCode, messageCode, message);
+	}
+}
