+++ src/main/java/codesquad/fineants/global/errors/exception/kis/RequestLimitExceededKisException.java
@@ -1,0 +1,7 @@
+package codesquad.fineants.global.errors.exception.kis;
+
+public class RequestLimitExceededKisException extends KisException {
+	public RequestLimitExceededKisException(String returnCode, String messageCode, String message) {
+		super(returnCode, messageCode, message);
+	}
+}
