+++ src/main/java/codesquad/fineants/global/errors/exception/kis/TokenIssuanceRetryLaterKisException.java
@@ -1,0 +1,7 @@
+package codesquad.fineants.global.errors.exception.kis;
+
+public class TokenIssuanceRetryLaterKisException extends KisException {
+	public TokenIssuanceRetryLaterKisException(String returnCode, String messageCode, String message) {
+		super(returnCode, messageCode, message);
+	}
+}
