+++ src/test/java/codesquad/fineants/domain/kis/client/KisClientTest.java
@@ -39,7 +39,7 @@ import codesquad.fineants.domain.kis.properties.KisProperties;
 import codesquad.fineants.domain.kis.properties.KisTrIdProperties;
 import codesquad.fineants.domain.kis.repository.KisAccessTokenRepository;
 import codesquad.fineants.domain.kis.service.KisAccessTokenRedisService;
-import codesquad.fineants.global.errors.exception.KisException;
+import codesquad.fineants.global.errors.exception.kis.KisException;
 import codesquad.fineants.global.util.ObjectMapperUtil;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
